Relocate side content
