+++ src/commons/assessmentWorkspace/AssessmentWorkspace.tsx
@@ -58,11 +58,11 @@ import { Position } from '../editor/EditorTypes';
 import Markdown from '../Markdown';
 import { MobileSideContentProps } from '../mobileWorkspace/mobileSideContent/MobileSideContent';
 import MobileWorkspace, { MobileWorkspaceProps } from '../mobileWorkspace/MobileWorkspace';
+import SideContentAutograder from '../sideContent/content/SideContentAutograder';
+import SideContentContestLeaderboard from '../sideContent/content/SideContentContestLeaderboard';
+import SideContentContestVotingContainer from '../sideContent/content/SideContentContestVotingContainer';
+import SideContentToneMatrix from '../sideContent/content/SideContentToneMatrix';
 import { SideContentProps } from '../sideContent/SideContent';
-import SideContentAutograder from '../sideContent/SideContentAutograder';
-import SideContentContestLeaderboard from '../sideContent/SideContentContestLeaderboard';
-import SideContentContestVotingContainer from '../sideContent/SideContentContestVotingContainer';
-import SideContentToneMatrix from '../sideContent/SideContentToneMatrix';
 import { SideContentTab, SideContentType } from '../sideContent/SideContentTypes';
 import Constants from '../utils/Constants';
 import { useResponsive, useTypedSelector } from '../utils/Hooks';
