+++ src/commons/sideContent/__tests__/SideContentAutograder.tsx
@@ -4,7 +4,7 @@ import { shallowRender } from 'src/commons/utils/TestUtils';
 
 import { AutogradingResult, Testcase, TestcaseTypes } from '../../assessment/AssessmentTypes';
 import { mockGrading } from '../../mocks/GradingMocks';
-import SideContentAutograder, { SideContentAutograderProps } from '../SideContentAutograder';
+import SideContentAutograder, { SideContentAutograderProps } from '../content/SideContentAutograder';
 
 const mockErrors: SourceError[] = [
   {
