+++ src/commons/sideContent/__tests__/SideContentContestLeaderboard.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { shallowRender } from 'src/commons/utils/TestUtils';
 
-import SideContentContestLeaderboard from '../SideContentContestLeaderboard';
+import SideContentContestLeaderboard from '../content/SideContentContestLeaderboard';
 
 const mockLeaderboardEntries = [
   {
