+++ src/commons/sideContent/content/SideContentHtmlDisplay.tsx
