+++ src/commons/sideContent/content/SideContentLeaderboardCard.tsx
@@ -2,7 +2,7 @@ import { Card, Classes, Elevation, Pre } from '@blueprintjs/core';
 import classNames from 'classnames';
 import React from 'react';
 
-import { ContestEntry } from '../assessment/AssessmentTypes';
+import { ContestEntry } from '../../assessment/AssessmentTypes';
 
 type SideContentLeaderboardCardProps = DispatchProps & StateProps;
